Run cleanup to get rid of all unnecessary casts
Ostensibly to remove casts to IVector sub-interfaces

+++ etomica-apps/src/main/java/etomica/kmc/IntegratorKMC.java
@@ -276,7 +276,7 @@ public class IntegratorKMC extends IntegratorBox{
     private double truncate(double numA, int digits){
         digits = (int)Math.pow(10,digits);
         numA = (long)(digits*numA);
-        numA = (double)(numA/digits);
+        numA = numA/digits;
         return numA;
     }
     public boolean checkMin(){
